Move rollback cmd to cmd/rollback
